day 6 part 1 complete
This was a fun usage of a pipeline. I fed in all possible coordinate pairs into the pipeline, and then ran multiple goroutines to crunch numbers and determine the closest coordinates to a given point. a final pipeline stage aggregates the data and outputs a result.

The final stage disqualifies coordinates if they either are on the edge of the bounding box or a location on the end of the bounding box reports the coordinate as its closest. In both cases, the area of points closest to the coordinate will be infinite.

Fun fact, my original solution was off by one. Remember, **the coordinate itself is part of the area**!