day 6 part 2 complete
this phase used the same point generator from the first section. The stage that consumes the point generator blindly adds up the distances to all points, then writes to the channel if the total is strictly less than the cutoff. Note that the cutoff itself is non-inclusive. This bit me at first!

It's nice to see this goroutine method work so well. Is it more efficient than just doing all the calculation in a single iteration? no. But darn it, it's pretty fun.